+++ src/vortenix_newsletter/providers/email/__init__.py
@@ -1,4 +1,9 @@
 from .console_provider import ConsoleEmailProvider
-from .factory import configured_recipients, create_email_provider
+from .factory import configured_recipients, create_email_provider, newsletter_recipients
 
-__all__=["ConsoleEmailProvider", "configured_recipients", "create_email_provider"]
+__all__=[
+    "ConsoleEmailProvider",
+    "configured_recipients",
+    "create_email_provider",
+    "newsletter_recipients",
+]